Move utils to root Discord directory and better organize utils
