+++ Discord/utils/dates.js
@@ -1,0 +1,9 @@
+const services = {};
+
+services.getDate = async () => {
+    let date = new Date();
+    let options = { timezone: 'EST', weekday: 'long', day: 'numeric', month: 'long', hour: 'numeric', minute: 'numeric' };
+    return `${date.toLocaleString('en-US', options)} EST`;
+};
+
+module.exports = services;
