+++ Discord/utils/fortunes.js
@@ -1,0 +1,18 @@
+const services = {};
+
+services.defaultFortunes = [
+    "Yes",
+    "No",
+    "Maybe",
+    "If you believe hard enough",
+    "Try asking again",
+    "Sure",
+    "Fair Enough",
+    "Please stop",
+    "Incorrect",
+    "You got it",
+    "Mhm",
+    "這都是中文的。當然"
+];
+
+module.exports = services;
