+++ Discord/utils/index.js
@@ -1,0 +1,26 @@
+const arrays = require("./arrays");
+const dates = require("./dates");
+const fortunes = require("./fortunes");
+const pagination = require("./pagination");
+const playingCards = require("./playingCards");
+const quotes = require("./quotes");
+const rankUtils = require("./ranks");
+const strings = require("./strings");
+const time = require("./time");
+const voiceConnection = require("./voiceConnection");
+const youtube = require("./youtube");
+
+
+module.exports = {
+    arrays,
+    dates,
+    fortunes,
+    pagination,
+    playingCards,
+    quotes,
+    rankUtils,
+    strings,
+    time,
+    voiceConnection,
+    youtube
+};
