+++ Discord/utils/quotes.js
@@ -1,0 +1,23 @@
+const services = {};
+
+services.thanosQuotes = [
+    "Fun isn’t something one considers when balancing the universe. But this… does put a smile on my face.",
+    "This… does put a smile on my face.",
+    "Perfectly balanced, as all things should be.",
+    "The end is near.",
+    "You should have gone for the head.",
+    `I know what it’s like to lose. To feel so desperately that you’re right, yet to fail nonetheless. Dread it. Run from it. Destiny still arrives. Or should I say, I have.`,
+    `Going to bed hungry. Scrounging for scraps. Your planet was on the brink of collapse. I was the one who stopped that. You know what’s happened since then? The children born have known nothing but full bellies and clear skies. It’s a paradise.`,
+    "I ignored my destiny once, I can not do that again. Even for you. I’m sorry Little one.",
+    "Your optimism is misplaced, Asgardian.",
+    "The hardest choices require the strongest wills.",
+    "You should choose your words wisely",
+    "The work is done. I won. What I’m about to do, I’m gonna enjoy it. Very, very much!",
+    "I am… inevitable.",
+    "You could not live with your own failure, and where did that bring you? Back to me.",
+    "I don’t even know who you are.",
+    "You’re strong. But I could snap my fingers, and you’d all cease to exist.",
+    "You have my respect, Stark. When I’m done, half of humanity will still be alive. I hope they remember you."
+];
+
+module.exports = services;
